- qt warnings

+++ copasi/UI/CQTSSAWidget.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2019 - 2024 by Pedro Mendes, Rector and Visitors of the
+// Copyright (C) 2019 - 2025 by Pedro Mendes, Rector and Visitors of the
 // University of Virginia, University of Heidelberg, and University
 // of Connecticut School of Medicine.
 // All rights reserved.
@@ -175,7 +175,7 @@ bool CQTSSAWidget::saveTaskProtected()
     }
   else if (pTssaProblem->getStepNumber() != mpEditIntervals->text().toULong())
     {
-      pTssaProblem->setStepNumber(mpEditIntervals->text().toLong());
+      pTssaProblem->setStepNumber(mpEditIntervals->text().toInt());
       mChanged = true;
     }
